refactor(pfb): extract barcode preview into FieldPreviewBarcode
